refactor(search): split filter/location providers into sibling file (#727)
Second file off the #727 hit list. `search_provider.dart` was 418
LOC; dropping 55 LOC of small helper providers into a sibling file
brings the main file to 363 LOC. Pure file split — no behaviour
change, same pattern as PR #794 (favorites).

- `search_provider.dart` (363 LOC) keeps the big `SearchState`
  notifier only
- `search_filters_provider.dart` (65 LOC) owns `SearchLocation`,
  `SelectedFuelType`, `SearchRadius`, and the `fuelStations`
  convenience accessor — all short, orthogonal helpers

`export 'search_filters_provider.dart';` at the top of
`search_provider.dart` keeps every existing import working.
`SearchState` additionally gets a direct import because it reads
`searchLocationProvider` internally (re-exports don't surface
symbols to the declaring file itself).

All 580 search tests pass; 0 analyze issues.

Part of #727 (2 of 4 files done — sync_service.dart and
add_fill_up_screen.dart still pending).
